Split into upstream-aligned files to ease syncing with libphonenumber
Reorganize the package so each file mirrors a class/module in upstream
google/libphonenumber, in upstream source order, so that a change to an
upstream file maps onto the same-named Go file.

This is pure code movement: no public API or behavior changes.

- phonenumbers.go      -> phonenumberutil.go   (PhoneNumberUtil)
- shortnumber_info.go  -> shortnumberinfo.go   (ShortNumberInfo)
- matcher.go           -> phonenumbermatcher.go
- extract enums.go, errors.go, metadatasource.go, prefixmapper.go,
  timezone.go, carrier.go, geocoding.go out of the 3.5k-line monolith

Each mirrored file carries a header pinning its upstream source and the
version it was last reconciled against.

The package init() trigger stays in phonenumberutil.go on purpose: Go runs
init() functions in lexical filename order, and metadata loading depends on
the generated protobuf types being registered by phonemetadata.pb.go /
phonenumber.pb.go first. That filename sorts after both, preserving the
ordering the original phonenumbers.go relied on.

Adds docs/2.0-restructure.md describing the file-split/method-order plan.
