+++ shortnumberinfo.go
