+++ shortnumberinfo_test.go
